Extract code style generating to separate project
