+++ Sources/Kysect.Configuin.CodeStyleDoc/ICodeStyleElement.cs
@@ -1,0 +1,5 @@
+﻿namespace Kysect.Configuin.CodeStyleDoc;
+
+public interface ICodeStyleElement
+{
+}
