+++ Sources/Kysect.Configuin.CodeStyleDoc/ICodeStyleFormatter.cs
@@ -1,0 +1,8 @@
+﻿using Kysect.Configuin.CodeStyleDoc.Models;
+
+namespace Kysect.Configuin.CodeStyleDoc;
+
+public interface ICodeStyleFormatter
+{
+    string Format(CodeStyle codeStyle);
+}
