+++ Sources/Kysect.Configuin.CodeStyleDoc/ICodeStyleWriter.cs
@@ -1,0 +1,8 @@
+﻿using Kysect.Configuin.CodeStyleDoc.Models;
+
+namespace Kysect.Configuin.CodeStyleDoc;
+
+public interface ICodeStyleWriter
+{
+    void Write(CodeStyle codeStyle);
+}
