+++ Sources/Kysect.Configuin.CodeStyleDoc/Markdown/MarkdownCodeStyleWriter.cs
@@ -1,7 +1,7 @@
-﻿using Kysect.Configuin.Core.CodeStyleGeneration.Models;
+﻿using Kysect.Configuin.CodeStyleDoc.Models;
 using Microsoft.Extensions.Logging;
 
-namespace Kysect.Configuin.Core.CodeStyleGeneration.Markdown;
+namespace Kysect.Configuin.CodeStyleDoc.Markdown;
 
 public class MarkdownCodeStyleWriter : ICodeStyleWriter
 {
