+++ Sources/Kysect.Configuin.CodeStyleDoc/Models/CodeStyleRoslynQualityRuleConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.Configuin.RoslynModels;
 
-namespace Kysect.Configuin.Core.CodeStyleGeneration.Models;
+namespace Kysect.Configuin.CodeStyleDoc.Models;
 
 public record CodeStyleRoslynQualityRuleConfiguration(
     RoslynQualityRule Rule,
