+++ Sources/Kysect.Configuin.CodeStyleDoc/Models/CodeStyleRoslynStyleRuleConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.Configuin.RoslynModels;
 
-namespace Kysect.Configuin.Core.CodeStyleGeneration.Models;
+namespace Kysect.Configuin.CodeStyleDoc.Models;
 
 public record CodeStyleRoslynStyleRuleConfiguration(
         RoslynStyleRule Rule,
