+++ Sources/Kysect.Configuin.ConfigurationRoot/ConfiguinCommands.cs
@@ -1,6 +1,6 @@
-﻿using Kysect.Configuin.ConfigurationRoot.Configuration;
-using Kysect.Configuin.Core.CodeStyleGeneration;
-using Kysect.Configuin.Core.CodeStyleGeneration.Models;
+﻿using Kysect.Configuin.CodeStyleDoc;
+using Kysect.Configuin.CodeStyleDoc.Models;
+using Kysect.Configuin.ConfigurationRoot.Configuration;
 using Kysect.Configuin.DotnetFormatIntegration;
 using Kysect.Configuin.EditorConfig;
 using Kysect.Configuin.MsLearn;
