--- Sources/Kysect.Configuin.Core/CodeStyleGeneration/ICodeStyleElement.cs
@@ -1,5 +1,0 @@
-﻿namespace Kysect.Configuin.Core.CodeStyleGeneration;
-
-public interface ICodeStyleElement
-{
-}
