--- Sources/Kysect.Configuin.Core/CodeStyleGeneration/ICodeStyleFormatter.cs
@@ -1,8 +1,0 @@
-﻿using Kysect.Configuin.Core.CodeStyleGeneration.Models;
-
-namespace Kysect.Configuin.Core.CodeStyleGeneration;
-
-public interface ICodeStyleFormatter
-{
-    string Format(CodeStyle codeStyle);
-}
