--- Sources/Kysect.Configuin.Core/CodeStyleGeneration/ICodeStyleWriter.cs
@@ -1,8 +1,0 @@
-﻿using Kysect.Configuin.Core.CodeStyleGeneration.Models;
-
-namespace Kysect.Configuin.Core.CodeStyleGeneration;
-
-public interface ICodeStyleWriter
-{
-    void Write(CodeStyle codeStyle);
-}
