+++ Sources/Kysect.Configuin.sln
@@ -23,6 +23,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Configuin.EditorConf
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Configuin.ConfigurationRoot", "Kysect.Configuin.ConfigurationRoot\Kysect.Configuin.ConfigurationRoot.csproj", "{71590EE8-BD16-4E45-8299-4D1CA3B0EC42}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Configuin.CodeStyleDoc", "Kysect.Configuin.CodeStyleDoc\Kysect.Configuin.CodeStyleDoc.csproj", "{372CD01F-754F-4D95-8F59-F08F9BC645A6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -69,6 +71,10 @@ Global
 		{71590EE8-BD16-4E45-8299-4D1CA3B0EC42}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{71590EE8-BD16-4E45-8299-4D1CA3B0EC42}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{71590EE8-BD16-4E45-8299-4D1CA3B0EC42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{372CD01F-754F-4D95-8F59-F08F9BC645A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{372CD01F-754F-4D95-8F59-F08F9BC645A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{372CD01F-754F-4D95-8F59-F08F9BC645A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{372CD01F-754F-4D95-8F59-F08F9BC645A6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
